Moves the 'cmd' package to the project root.
Moves the 'cmd' package from the './internal' directory to the project
root. In addition, the 'just build' command has been extrapolated into a
'just _build' command which allows developers to run builds without the
'just web/build' dependency.
