+++ cmd/connect-client/commands/debug.go
