+++ cmd/connect-client/commands/publish.go
