+++ cmd/connect-client/commands/version.go
